+++ src/test/groovy/ru/vyarus/gradle/plugin/python/VenvFromVenvCreationTest.groovy
@@ -3,6 +3,7 @@ package ru.vyarus.gradle.plugin.python
 import org.gradle.api.Project
 import ru.vyarus.gradle.plugin.python.cmd.Pip
 import ru.vyarus.gradle.plugin.python.cmd.Virtualenv
+import ru.vyarus.gradle.plugin.python.util.PythonExecutionFailed
 
 /**
  * @author Vyacheslav Rusakov
@@ -21,14 +22,14 @@ class VenvFromVenvCreationTest extends AbstractTest {
         Virtualenv env2 = new Virtualenv(project, env.pythonPath, null, "second")
         env2.python.extraArgs('-v') // enable logs
         Pip pip = new Pip(project, env.pythonPath, null, false)
-        pip.install(env2.name)
+        pip.install(env2.name + "==16.7.9")
         env2.createPythonOnly()
 
         when: "validating pip in second environment"
         Pip pip2 = new Pip(project, env2.pythonPath, null, false)
-
-        then: "INCORRECT: pip exists, but it shouldn't"
         println pip2.version
 
+        then: "pip not exists"
+        thrown(PythonExecutionFailed)
     }
 }